HUGE UPDATE & DOCS

+++ Crystal/src/Crystal/Renderer/Buffer.h
@@ -31,7 +31,7 @@ namespace Crystal {
 	struct BufferElement
 	{
 		std::string Name;
-		uint32_t Offset;
+		size_t Offset;
 		uint32_t Size;
 		ShaderDataType Type;
 		bool Normalized;
